Add first version of project template
Create Python project template with packaging and testing support.

Packaging:
With `setuptools` using `pyproject.toml` in and `setup.cfg`.
Transitioning to a project fully on `pyproject.toml` would require
either switching to Poetry or wait for `setuptools` to addapt.

Testing:
With Pytest, MyPy, Tox, coverage and generation of HTML reports.
Coverage is currently buggy but I have spent too much time in this for
the moment.